Mesh_3: Fix other conversion warnings

+++ Mesh_3/include/CGAL/Mesh_3/vertex_perturbation.h
@@ -342,7 +342,7 @@ public:
   double time() const { return timer().time(); }
   int total_counter() const { return total_counter_ + counter(); }
   std::size_t total_time() const 
-  { return static_cast<std::size_t>(total_time_ + 1000*time()); }
+  { return static_cast<std::size_t>(double(total_time_) + 1000*time()); }
   virtual std::string perturbation_name() const = 0;
 private:
   CGAL::Timer &timer() const
